Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  dd593786a262c93272ef21c5f3d170313e253e02
Honor language=xx option in parserTests.txt

This patch implements the language=xx option in parserTests for
wt2html.  This may expose some failures in l10n.  In particular, the
tests are now run in the 'en' prefix by default (as thay are in the
PHP parser tests), and the linktrail and linkprefix for 'en' become
significant.  A bug in the WikitextSerializer is also
exposed in the "Simple category in language variants" test case, which
doesn't properly normalize the category link name when the locale is
set.

Change-Id: Ibb925ecdfe848b43e46ff97b29c5bddcdb02e9c8