Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  363ed89559e0c155e4a0f0802ffd0142611f494a
Fix photo upload button not working after error

When an error occured when adding a lead photo to an article, the photo
upload button no longer worked and the user was be unable to
retry.

This was caused by jQuery's .remove() unbinding all DOM events, .detach()
should be used instead.

Change-Id: Iba136eeb77ec4eea0b01eb0ded52e813471b50d2