Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  81a8cbdf8204e749410a72277700b493e7975915
User-friendly boolean options for parserTests.js.

Allow --debug=no and --debug=false to mean the same thing as --no-debug.
This saves some hair (why doesn't "--use_source=false" do anything?).

(I'd actually expect the 'optimist' package to do this for us when we
specify that an option is a boolean option... but sigh.)

Change-Id: Icd89a0fb59e3a19cf88b49e24b67eaebf3eaadea