Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  546cc44a0658cf85319f16e8d941027a4c645020
(bug 44934) Remove link to Special:PropertyDisambiguation in new property page

We still need a need for some better way to search for properties. Once we have something
like that, then it would be nice to link to that from Special:NewProperty.  Until then,
it's silly to link to a special page that does not exist.

Change-Id: I6313038bcc9f098d562f277549b5382b1bba4cde