Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  263af1b128861035ed3a8d7d503defaaacd66bfb
Moving "entitysearch" to experimental features

Replacing the native search box with the entity selector requires some adjustments
to be done in the entity selector widget. Until these are not done, the entity
selector should not replace the search box since it will behave very different than
a regular search box.

Change-Id: I37786e885f48f0d4f80025a8835f10512ff0a7f4