Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  202f51b975518c2d91b689291f6fec96a550e3ff
(Bug 45352, Bug 45353) Add ItemByTitle and ItemDisambiguation to the special pages list

Changes so ItemDisambiguation and ItemByTitle can be added to the list
of special pages. Also some minor changes to the messages used for those
pages.

There are also a name change in the list to ItemBySitelink, we are not
very good at following our own naming scheme. We should really try to
follow our glossary.

Change-Id: Id197c49dd52c482eb162f56522e5f0db0256b90a