Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  dfae82d40a5cf885a8921766b81b04afb7639fd0
Standardize the hamburger and remove isOverlay (#375)

isOverlay is not needed anymore because all pages are supposed to have
the hamburger. Articles have a searchbox, special pages have a heading.

Change-Id: Ie43b568221a067a0eaf2923bfc3a779bb8ca21ca

Remove noMargins and the need for addArticleClass()

Also set isOverlay to false (it was adding a .content div to many
special pages) and prepare for its removal.

Change-Id: Ie77849e636a4d7420cabae88ef7fffd9a6f85b04