Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  4db4bcc8e13f37b45cee6f8220e347bf3ebe5a35
Change the way main page is detected

Looking for "Today's featured article" is Wikipedia-specific. Looking
for the div tag is not very elegant but it seems to be the only thing
that always distinguishes the main page from other pages.

Change-Id: Ib9315f076b4f3f0c894dfa465baaacdf616f7b9d

Force mobile view when using local_browser

Local dev instances of MediaWiki usually don't have a separate domain
set up for the mobile view, but it can be triggered by setting a
mf_mobileFormat cookie.

Change-Id: I09e7fe4b70d4aad25161f0c9ca878e84e27749f3