Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  e457ee70b4b8ceba04dc6c9cef35f45b2b1974a9
Remove Git SHA1 logging functionality

It's a bit out-of-scope for EventLogging to be mucking about the call stack and
file paths. My initial enthusiasm for the idea has waned. If we got serious
about adding this functionality, we should do it the Right Way: by improving
GitInfo.php in core (it should be able to search up a path for a .git dir, for
example). If we went that route and reached a point at which adding this
information to server-side events could be done simply and reliably by invoking
functionality in core, I'd be interested in re-introducing it. But it's just
too much stuff to pile into EventLogging.

Change-Id: I6fb205830eb21a05a8f5d4594ad875468cb78377