Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7a41726b227b5f2956f46ed2de88546ceaf7f5ed
Introduction of wb.Entity and related constructors and its usage in our code

Constructors we need for first step towards proper entity store interface which we can use in
entityview widget and others to get rid of global revision and entity stores. This patch set will
not make any real use of any of that new code in existing code yet.

* wb.Entity, wb.Item and wb.Property implementations which will replace loose array structures we
  are dealing with currently.
* initial implementation of wb.store.FetchedContent which represents some wiki content plus meta
  data such as wiki page title and fetched revision.
* new module "wikibase.serialization" which holds serialization logic to unserialize
  wb.Entity and wb.store.FetchedContent from JSON.

Change-Id: I1c8912d8e82ff1b563a6e8861f73bfb223caa696