Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  74e13e4ac9be566e438009ba246c680f075155aa
Use mock API to test parser extensions

Start a mock API from parserTests, and use it instead of any real API.

This seems to change *a lot* in our tests output, including at least one new
failure. But that failure was there before, we just didn't have any tests to
expose it. Namely, "Simple category in language variants (wt2wt)".

It's also *fixing* a lot of tests. But no matter. The point is it seems to work
as expected.

We now also have an --extensions option for parse.js. Use it to add acceptable
extensions - they'll be expanded on enwiki (or whatever wiki you have configured
via --prefix or --apiURL), so any extensions you add should be installed
there, or they'll just be returned as escaped text.

Bug: 45440
Change-Id: Ie63cb796d098947443616c8bb02ebdea901815c0