Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d88b34a0bccd5a31e138b85da08744ce7e3a6061
(bug 44710) Don't display (none) in case we don't have langlinks

The "in other languages" box is now getting shown by the
linkItem.js by the time it's ready to be used (and to
add new items).

This required adding a new CSS module with position=top
as we can't hide this in JS (as JS is only run after the
element is already visible) nor in the existing CSS (there's
no way we can let it know whether there are langlinks or
not).

The message 'wikibase-nolanglinks' became obsolete by this.

Change-Id: I45790dfb959e7b79fb1581fe16bdb25e99c90102