Updated mediawiki/extensions Project: mediawiki/extensions/OAI  924bdfb83ee07623ac1eee3a39f2899b5f23cea2
Revert "bug 45266; needs parallel changes to lucene_search_2"

This reverts commit 0e035870dc693d1d6ad82ced83c1d6adbdf9d361

I'm pretty sure this will instantly break all indexing if it is
deployed. OAIHarvester.hasMore() will always be true, so
getMoreRecords() will continue to loop until 50 records appear, which
may not seem so bad until you realise that this is done for each wiki.
It might take a while before 50 new records appear on aawiki.

The patch to hasMore() would fix it, but this update will go out in
the normal release cycle whereas the Lucene patch will not. And there
are presumably other OAI clients.

It's an OAI protocol break, which I think is OK as long as it is
opt-in. Either the client should opt in with a special URL parameter,
or the server should provide a backwards-compatible response format
which uses some tag name other than resumptionToken, i.e. a tag name
which doesn't have a particular meaning defined by the OAI-PMH
protocol specification.

Change-Id: I0f1be5afec98fbfcf93471809b633a55e61a8279