Updated mediawiki/extensions Project: mediawiki/extensions/TwoFactorAuthentication  dbe78e6fe198e948b17f8ac37fc320ecdc29725d
Added separate pages for authentication; i18n typo.

Added functionality so that the user enters the
OTP on a separate page rather than on the login
form (mainly for UX so users without a token are
not confused by the field).

Also fixed a typo in the extension messages and
added a missing Qqq message.

Change-Id: If0b48654cab8d1ded4410fbf894ed0eaa1f71bef