Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d5c48863da58ea23f574548f753713617fbeb90c
Add dt/dd msgs for of the datatypes in Special:ListDatatypes

These messages are used at the "List all datatypes" -special page.
In those cases some datatypes exists that can be added as configuration,
the messages can be set upp manually in as local system messages.

Patchset 9: Only a rebase

Change-Id: I66ed2cf6e82d50fc0904c89d37e09daa88a2a6ad