Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  538fcca2c469d2e5d8d62fd8a03c9c190235e001
Back to weird title

In some part of the extension we are using words that simply don't
fit what we are doing, but some of the weird wording has stuck
somehow. In ItemByTitle we are really talking about a sitelink to
an external site, with a pair of site id and a title on the external
site, not about a single title. The last would be a local page.

Change-Id: I40f69948036cec605dc0aa342084d884a003dd50