Updated mediawiki/extensions Project: mediawiki/extensions/OpenID  e4793fb7c93cf993594bb753e043d4c8037e21e6
fix a regression: User profile userinfo was not updated from OpenID provider.

This fix corrects a regression due to a recent change in the message texts.

Changes in the message texts are now backported to the corresponding database
key names, which must comprise a "-" sign between prefix and the key like in
"openid-update-userinfo-on-login-email" where "-" and "email" are now
correctly used.

This fix regards MediaWiki when it is used as OpenID consumer and when
a user logs in with an OpenID eg. from Google.

It repairs the mechanism of setting and checking the user properties on the database
which determine what part of the userinfo (nickname, email, ...) the users allows
to be updated automatically from OpenID provider when loggin in with OpenID.

Change-Id: If49a33d258e2a62effb0c59b6a010d60719f3fb7