Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  361cbd620bbf78748a51a4c3c7f11ecdc36231a9
(bug 45159) Introducing snaklistview widget

The snaklistview widget will manage multiple snaks by using a listview widget.
The snaklistview's first use is in the list of references. This change set is
only the first iteration that resolves the referenceview's inheritance by
defining the new snaklistview widget as parent instead of the claimview widget.

Change-Id: I25f64517f5c7d98b330f4d6a5620b7f62280de4d