Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  fab7dba0d01833bc3aeeb1b16994141c4273c72c
(bug 45159) Splitting snaklistview from referenceview

This change set separates the responsibilities of the snaklistview widget from the ones
of the referenceviewwidget. Now, the snaklistview is actually manages just a list of snaks
while the referenceview just adds some additional functionality regarding the handling of
references. Furthermore, the statementview manages removing the referenceviews.

Change-Id: Id677e6a6387661f00945d6b5dcaa8603705530cb