Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5f0ff02711bc42baf8829ae3840a992d36f4a976
Moving "add" toolbar initialization from snaklistview to referenceview

The referenceview should control the toolbar to add new snaks to its list. This allows
adapting the "add" toolbbar to the toolbar controller specifically for the referenceview
which will be done in another change set.
Along with this change, the "add" button for adding snaks to a reference will be disabled
as long as the list of snaks is not valid.

Change-Id: I1e9841e1683e6d503be1ffd654fd97f62b93f7aa