Updated mediawiki/extensions Project: mediawiki/extensions/Math  20192a1ab87d092963af732fdd8c95e5e60509a8
Do proper pre-pack processing for $this->hash.

Output hash calcuation sometimes fails because of Texvc failure for
example, and $this->hash remains ''. At this time inserting null as
outputhash into database is improper, because that column is NOT NULL.

With this patch an empty string '' is inserted instead, and the if-check
is modified to avoid cases where ->hash exists but evaluates as null
(for example when it's '0000000000...' though it's extremely rare).

Change-Id: I852859f4b151b777c11b743faaed61dfc2c029a7