Updated mediawiki/extensions Project: mediawiki/extensions/AccountAudit  c9ae63b1aaf397916d82b2bc243559a2b9b65e4d
Initial commit of Extension:AccountAudit

The extension adds a table for tracking the timestamp of the most recent
login for a user.  This will allow us to determine the (in)activity of a
user and thusly whether or not we can unilaterally rename the user in the
event of clashing local accounts during unification.

Change-Id: I163682c55031e77bba02d67cd62839f9541adc99