Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  b58ee1da94b9666092e01127a4d58ae01c4e5494
Use the new limit report hooks

Change Ie065c7b5 added an option to show profiling data at the bottom of
preview pages, and with it new hooks to gather and format this data in a
more structured way than is possible with ParserLimitReport. This change
adds support for the new hooks.

Depends-On: I7799616a602d90e1b8d3f0ece35811ca387bade7
Change-Id: Idffd2d78f9a0217c99c07cbbfc844d6daf0172f7