Updated mediawiki/extensions Project: mediawiki/extensions/OpenID  e28970a0e1c77e330994f886977885f34e9188b8
Show OpenIDIdentifier page with user information only for logged-in user

On the OpenIDIdentifier page, a text says that this is the identifier
for user xyz but,

new: show only the name when xyz is visiting it.

Never show a name of a different user.

But show an "invalid identifier" message if a user with a corresponding
User ID is not known.

Made the SpecialOpenIDIdentifier page an _unlisted_ SpecialPage, as it makes
no sense to list an unspecific SpecialOpenIDIdentifier among the other
Special Pages. It would only make sense if we can list it as
SpecialOpenIDIdentifier/id , but this is not possible.

Users can see their OpenIDs already on two places:

i)  below their user page name, as a page sub-title and
ii) in the OpenID Preference tab.

Change-Id: I66b4d2e43729c1f867f3762e651b5fc081f2bac1