Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  db90cdd194f74c3264af00df865672b45acb49f6
(Bug 44536, Bug 44968) Use SiteLink instead of Title

This changeset uses the local names on the namesaces, with the
additional overhead that creates. To avoid some of the trouble
that creates the special page uses a short but important
setSquidMaxage(10).

Patchset 14: Changes according to Daniels wishes
Patchset 15: Rebased
Patchset 16: Some changes to trim and normalize both site and page

Change-Id: I1176f839579f936cecf177310431c1192a74dad8