Updated mediawiki/extensions Project: mediawiki/extensions/Agora  5736fa593f8e91901dfb93d666dfd1c72f5b3dd7
Generate varying CSS for Vector skin vs. default.

This implements Trevor Parscal's suggestion to use skinStyles to send
different CSS for the mw-ui-* styles to the browser according to the
user's skin.

The CSS uses the new naming mediawiki.ui.{vector,default}.css; the
source .sccs files use this naming too which simplifies the Makefile.
The generated CSS is identical to the previous ext.agora.css (except for
an '@charset "UTF-8";' in my compass generation).

The next step is to modify mediawiki.ui.default.scss to, e.g., not
include the button colors.

(A separate patch will incorporate the CSS into core MediaWiki as part
of the new-look login form.)

Change-Id: Ie7b4af704a0a158bdaf7e908941bd00d1e2fd50c