Updated mediawiki/extensions Project: mediawiki/extensions/Translate  c0702ab95274c8abc281c8ffc3dd871240c9ab26
TUX broken if ULS is not enabled

Made TUX work even if ULS is disabled be explicitly adding uls
resource loader module to dependencies.

This does not add the language selector to the page, but it
enables input methods and webfonts and also ui language setting
which does not work if used. More work in ULS is needed to
only load the necessary things for language selection.

Bug: 45893
Change-Id: I307c2fe1924c92eeee22541ffe8db9edd92ed455
Mingle-Story: 2716
Mingle-Task: 2787