Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  102ee1323325a6549f122189a528671a6367245b
Removing entity selector padding

The default entity selector padding influences the appearance of the entity selector
replacing the native search box which should use the browser's default padding.
If a certain entity selector padding is necessary, it should be set more specifically.
This will probably be referenced when reviewing the css for the statement section.

Change-Id: I2901ee031ff429ce69bb355a6e181043d7853e5c