Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  6faa00b956534ff85b2ed7f039ebc60304709bf6
Add Special:LoginHandshake workaround (#439)

This makes the login (and create account) page redirect the user to the
login handshake URL on their first login attempt.

It's enabled by setting $wgMFLoginHandshakeUrl to a URL of the
Special:LoginHandshake page we want to redirect to (in case of Commons
it would be http://commons.m.wikimedia.org/wiki/Special:LoginHandshake).

Special:LoginHandshake sets a 'loginHandshake' cookie (with no
particular content) which reassures mobile Safari (and future versions
of Firefox too, probably) that it can accept third party cookies from
this domain. As soon as the cookie is set, it redirects the user back to
the login form.

Change-Id: If45e29ffd0bb9c4c945f942951a8e482017610c8