Updated mediawiki/extensions Project: mediawiki/extensions/ZeroRatedMobileAccess  0bf768f20b2500dbbc7fa9b4373f64fe47458268
Reducing default memcaching of carrier preferences from 3600
seconds (60 minutes) to 900 seconds (15 minutes). After this
reduces the pain, we can consider refactoring to hit the local
endpoint to obtain the options.. Currently, the carrier prefs are
pulled remotely via an HTTP GET off of en.wikipedia instead of
with a straight wfMessage hit.. The three cases where the prefs
are cached rely upon self::getMaxAge(). self::getMaxAge()
is not invoked elsewhere, meaning this is isolated.

Change-Id: Ie86dc96fdd9e7cca482780cb1228d1394f07a4e9