Updated mediawiki/extensions Project: mediawiki/extensions/DismissableSiteNotice  cf419da8020c79cb45485a0d14f369de485031f2
Make use of the ResourceLoader

Instead of manually building inline JavaScript, which involved
using the now-deprecated Xml::escapeJsString(), use ResourceLoader
to load the minified JavaScript and CSS separately.

As part of the changes in the way HTML is built, I added a new
message "sitenotice_close-brackets". This should address bug 17089.

Bug: 17089
Change-Id: I069f7f70c2d2a8a0f403452d6ccef82f659418af