Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  880696ae5bc9aba86ba3bfc3edc1cb9ad1a7caaf
(bug 46905) Add "centralauthtoken" to API

Since third-party cookies (and therefore SUL) don't work with a growing
number of browsers, we need a non-cookie-based way for CentralAuth to
pass authentication data for CORS requests.

This adds a "centralauthtoken" parameter to the API "main" module, and a
new token type to action=tokens. The client requests a token from the
local wiki and then immediately uses it in the CORS request. The remote
wiki will then ignore all cookies and instead take the user data and
session ids from memcached using that token.

Bug: 46905
Change-Id: Ia3657a5f7c9f1f0035f2b244d56995a8ed8a9cc0