Updated mediawiki/extensions Project: mediawiki/extensions/Lingo  5161518bac2f80e6385581fac63b7d2681023782
bugfix (links with terms broken)

"In version 0.4.1 all links, where part of the link is a term, are corrupted - they do not appear as links but instead the wiki text is shown. With 0.3x it worked."

Using ParserBeforeTidy for MW pre1.20 now. This reintroduces the problem where text in nowiki tags is annotated. Can not help that.

Change-Id: I630e49ee6adb4f9d1d5a8f058401f4c311eb5432