Updated mediawiki/extensions Project: mediawiki/extensions/Translate  b3e1a58765da1c41bc2eabb75810b1856f0266f5
Readd formatMessageBeforeTable hook

The hook has a PHP variant in MessageTable.php, but that is not
replicated in TUX at the moment.

Includes documentation.

Change-Id: I1080132bf5c85d70722bddf8fc977db8c41dc3dd

Make it possible to use a custom selection of translation aids.

For both removals and additions, the list of TranslationAids
available must be set on group-by-group basis. Here, the existing
(static) method provides the base set to the MessageGroup and
MessageGroupOld base classes. Changes can then be made by
overriding the MessageGroup(Old) default set.

Custom removals are effected by allowing use of a new
UnsupportedTranslationAid class that always errors, ensuring
that handlers need only check for the presence of an error response
rather than also testing for the existence of a response at all.

Custom additions require the new identifier to be registered using
a repurposed TranslateTranslationAid hook in addition to the obvious
addition to the base list when overriding the base class.

Also, expose the group-type via Tux to allow for usable hook runs
(it is already exposed in the default interface).

_autoload.php and hooks.txt updated accordingly.

Change-Id: Ifb0a2d510487bb5f888930a824d227f1ae6d50b2