Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  a749b6008d464471f21d1a955f0da0756114c886
Remove the need for a button in sections

The fact we add an unnecessary button to the DOM has irked me for a
while. It is only there for historic reasons. This patchset removes it
and throws away lots of code in process

This does however make a slight change which will need to OKed by
design - the buttons are now always black (previously in closed state
it had opacity 0.5)

Change-Id: Ic501c8ccb65a72c85abbe386fe6504960b1665a2