Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  5ba437dcd27b4258a301c55239b6130665e9d673
Check that parserTests.txt is in sync with upstream copy.

Warn (but don't update) when running parserTests.js, because we might
have a good reason for using an unsynchronized copy --- for example, we
might be testing an unmerged patch, or working on a branch.

"npm test" will always update parserTests.txt and then run the tests.

Change-Id: Iaeb674bb3459658e11727130d1046041ae5d078d