Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  729a09b890fcc38a436245210b467862ccf9c0b2
(bug 46720) attempt PayPal localization

This is working for one-time donations, but currently does not for
recurring donations.  The API params seem to be correct, so I'm
assuming this is a bug or docbug on PP's end, for the moment.

There are really strange problems with the way we're forced to
specify locale, for example, they have full spanish translation,
but the interface will not appear in Spanish unless you happen to
be located in Spain (or falsify the "country=ES" param...)

Change-Id: I6afd7f57c6c2e7f7d26c18804c5eef13804404ef