Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaMaintenance  f24470e6487961b08d68ddf87d0752b810dc4e6f
(bug 27320) Add refreshMessageBlobs.php

The clearMessageBlobs.php script is a bit too heavy-handed: it
obliterates the entire ResourceLoader message caches for all wikis,
which can lead to a database stampede as user requests try to repopulate
the cache all at once.

refreshMessageBlobs.php takes a more nuanced approach: It iterates over
all the cache entries and updates only those where the message cdb has
changed. And it does this one resource at a time, rather than having
every client connection try to do it in parallel.

The downside is that this is a good bit slower.

Bug: 27320
Change-Id: I3b6ae12875f2f323210fdfba36c5c5d9183588e2