Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  30f597a1087f48b3aa8feebbfc1fbfeefae8494f
Let more attribs through in parsoid-only tests

We need e.g. property and class so we can make sure the new image tests
actually test what we need 'em to.

Only two parserTests changes:

1) "Table with broken attribute value quoting" now passes wt2html
because title/style attributes aren't stripped.

2) "Italics and bold: other quote tests: (3,2,3,3)" fails html2html -- it
previously failed wt2wt and html2wt (ie, we don't round-trip this cleanly)
and managed to pass html2html only by a fluke (we were stripping out
the nowiki <span> attribute it was adding).

The underlying issue with the failing test case is that we round-trip
   '''this is about ''foo'''s family'''
as
   '''this is about ''<nowiki>foo'</nowiki>''s family'''
which is conservative but the <nowiki> is not strictly necessary.

At any rate, we now accurately fail wt2wt, html2wt, and html2html on this
test case.

Co-authored-by: C. Scott Ananian <cscott@cscott.net>
Co-authored-by: Mark Holmquist <mtraceur@member.fsf.org>
Change-Id: Idc0de45d43fdecfd37c730baee10c8ee086e6fd6