Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  e2ac1507d5fa40ea1dca1cfb5aff2569a368858f
Updates to migrateAccount.php for performance.

Adding the option to pass a list of usernames with each username on
a new line to prevent having to initialise an instance of MW for each
username that we are migrating in hopes that this is the cause of the
script being so slow.

Also adding a wfWaitForSlaves() call since we have a good opportunity
to do so.

Change-Id: I382d5501fa1cf209b26db5a52b45b6d04c7f3c65