Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  66c272e353b855d5a899414686cbd80947034a39
Bug 47290: Fix search in non-js mode

Due to a combination of the negative right value and an empty
meaning no users on the simple version of the mobile site or with
javascript disabled can search.

This is bad!
This makes slight adjustments to form based on whether js is enabled
or not.

Change-Id: I1bcffd26e5770e3cec1865c0c4fd614b6835a905