Updated mediawiki/extensions Project: mediawiki/extensions/TranslationNotifications  b193d56aeb85c87c9347ee1bc59cecb95fa1083c
notifytranslators.js: Fix broken closure

This was doing the following:

    jQuery( document ).ready(
 	    function ( $, mw ) { .. }( jQuery, mediaWiki )
    );

Which does just this:

    ( function ( $, mw ) { .. }( jQuery, mediaWiki ) );

    jQuery( document ).ready( undefined );

Rendering the docucument-ready call useless.

It mangled a closure with a plain jQuery document-ready callback
(which never take "mediaWiki" as argument, obviously).

It worked fine in most cases since this module is loaded from
the bottom so the document is mostly ready.

Fixing. Ignore whitespace since block level was changed for most
of this file.

Change-Id: Ifffafd75799cc60c4f09792cb993ff54b2f875ad