Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  2bcc7d13384300063d7e88a090df1336ca1caa7b
Use RecentChange::getTitle in ExternalChangesLine

Title::newFromText will use the given namespace as default, when parsing
a title with a namespace at the begin. This method will not use the
default, instead used the namespace from the given text.
dewiki used some categories with for example file: at the begin, which
have a wrong link in the watchlist/recentchanges. Fixed by using
RecentChanges::getTitle to get the correct title. Another solution is to
use Title::makeTitle direct.

Bug: 46886
Change-Id: I0a58cd4a8d98b94ccde1b70d25a4a3650a7d8c08