Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  aef4d4982686db0059c7c48cd0191a2fbfde36e8
Handle all the Node.js dependencies the same way

Add jshint as a dependency to package.json. Enforce minimum versions
for jshint and autoless.

Now everytime a Node.js Makefile task is run, it checks if the
dependencies are up-to-date and the versions can be updated easily for
everyone working on MobileFrontend.

Change-Id: I68de83ffcd383bc6dbb28af6af8b407998097ddf