Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  d0ce8153695485f7fdba9ef71c37cfd2b761d74a
SMW\Test\PropertyStatisticsTableTest issue on Windows

I excluded the Datagroup from running tests in the past but now
I need those but PropertyStatisticsTableTest still complains
on Windows about pow( 2 , 31 ) being "MWException: The value to add
must be a positive integer" and since I don't have to time to research this
in detail, I added isWinOS() which checks if the test runs on a Win
platform and if so do pow( 2 , 30 ) instead.

Doing pow( 2 , 30 ) might not sufficiently serve the test purpose and
I do not know why Win is complaining about pow( 2 , 31 ) but I need the
group Datagroup tests to be run without failure.

Change-Id: Icf5243f37919f9784e884d5cb81faa023a2e2891