Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  018f6f7be0eb475e5099d20388a04592696bbead
Added IDs for entity page's sitelink and claim headings

This introduces the IDs "claims" and "sitelinks" for h2 headings of claims and sitelinks. These
should only be available for the "main" entity of a page, not by entities displayed on another
page than their actual article page.
Put TODOs in the related EntityView code since this is something we might want to consider when
refactoring it.
On the JavaScript this is already ensured since we are only adding the IDs from the code which
controls the instantiation of the entity related widgets on an Entity page.

Change-Id: I9f3a54ca2f807e680cbe9bdcbaa3f76bbf4c251f