Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  b2efb2d601af261b30e88b43121064f943db8a1f
Fix crasher on :fr:126e_méridien_ouest

This code is being run on:
  <meta property="mw:PageProp/categorydefaultsort" content="0 126">
which doesn't have a 'typeof' attribute, and so the match is crashing
with a null pointer dereference.

Turn the expression around. RegExp.test is more efficient, and it handles
a null argument with no problems.  Add start/end anchors to the regexp
while we're at it.

Change-Id: I11bc9fdc902a21ce0bc1c05f08d3b615d5483e2e