Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  4407a8d9adb60efab1b24123da92aed41455ef48
Don't use focus event for the search input

Use touchend and keydown instead. This way the search overlay is shown
only when someone taps the input or starts typing in it. The overlay is
not shown anymore when someone uses Tab or Shift+Tab to navigate the
page. Also fixes Kindle annoyances.

Bug: 47499
Change-Id: I99347dd90a20e796551a1ce6aed47278fcb40c0e