Updated mediawiki/extensions Project: mediawiki/extensions/Math  eaed42d8e42e798469fec1ad781c06ae87a0979c
tests: skip math table creation for SQLite DB

SQLite has global indices, not per table one.  When using the database
updater to create the `unitest_math` table, we would get an error
with a duplicate `math_inputhash` index which is created by the the
database installer.

This patch skips the testBasicCreateTable() test on sqlite.

Change-Id: I6da56d67d7af45f86cf6163cf73f5464505b43f3