Updated mediawiki/extensions Project: mediawiki/extensions/DataValues  0f7370d5e569037347b2a18624e8977fda183489
(bug 48145) Introduction of TimeValue data value type in JavaScript

Changed the backend's TimeValue precision numbers to match the one from the frontend since
the time.js's options and precision handling is quite a mess right now. For having a simple
version of the time value frontend running asap, this is the faster way to get there.

Change-Id: Id9db92258aa667cc625df1483b0dd5aaf8dd3c7b