Updated mediawiki/extensions Project: mediawiki/extensions/TorBlock  b7be2169909d1f809c92f3f5dfff79b28af6e9c4
Fixed classification of Tor nodes to only block exit nodes.

Switched Onionoo to use the detailed summary and to query
specifically for exit nodes so that only exit nodes are
blocked. Also, since the flags are now set in the URL itself,
the Onionoo response will be reduced to only relevant relays.

Bug: 47626
Change-Id: Ib15a9ab41ed2d3c2b6e39067e1bd9076a8b6888f