Updated mediawiki/extensions Project: mediawiki/extensions/Translate  ec8a3e71ce7ff7a1a035bc966e4e3b8a656b5448
Apply dir and lang to the target language selector trigger

Before this patch, if a language with a different directionality
is selected as the target language, its name may be shown incorrectly.

This patch applies the correct lang and dir attributes to the element
that contains the language name so it would be displayed correctly.

Change-Id: Ib830a5856ec41ea795fa738e5eb58c40c09ce7d7