Updated mediawiki/extensions Project: mediawiki/extensions/CodeEditor  b802deafbfc85a36b67c687b2af72ce02ede6400
Make CodeEditor cooperate with the experimental live preview

LivePreview takes the text from the textarea and sends it to the server
via an Ajax call to handle preview and diff. But CodeEditor hides that
textarea, and only updates it when the form is submitted or the "disable
CodeEditor" button is used.

Fortunately, LivePreview already provides a custom JavaScript event to
indicate when it is about to pull the textarea contents. So all we have
to do is listen for that and update it.

Bug: 47235
Change-Id: Iba1b727c55aabcd31afc06cec99618fcdce207d8