Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  553d6a4c8491e351e39e4fb4229104311b7a4b11
Better regexp for "typeof" attribute; use normalizeOut for parsoid-only tests.

1. Use consistent normalization for parsoid-only tests (normalizeOut for
both parsoid output and the output specified in the parserTests file,
although we normalize the DOM read in from parserTests before normalizing).

2. Match "typeof" attributes during normalization, even if they are not the
first attribute in their tag.

Five different parserTests newly pass.  There are two new failures, both due
to the fact that we accurately strip <span typeof="Object/Template"> now.
In the "Ugly nesting" test case, the typeof used to be hidden by another
attribute in front of it; aka "the output was always wrong".  In the
"confirming safety of fix" test case, an initial <span> in the template
content is used instead of a newly-added <span> wrapper, and so normalize
strips an attribute when it shouldn't.  Both of these will be fixed
in upstream parserTests.

Change-Id: I823e5515698764b5b7fbed4a1bd902f8233e71c7